requirements.txt
Flask==3.0.0
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-monitor==6.0.2
azure-monitor-query==1.2.0
gunicorn==21.2.0